refactored testing so that setup and teardown is derived from a base class and tests are split into classes by section of the system they test
